Add STRING default memory flags and handle a related special case
RT_STRING if used as a resource type directly via its numeric value leads to creating a STRING resource in the .res with the format of a user-defined resource. This is *not* how RT_STRING resource types are formatted, and the resulting .res file will likely cause anything trying to use it to explode (e.g. it causes Resource Hacker to error with 'Unable to read Resource Section').

Instead of emulating the Windows RC compiler and creating bogus .res files, resinator will make this particular edge case a hard compile error.