parse: Avoid incorrect lexing contexts by using lookaheads
The previous implementation would try to anticipate what type of token should come next (i.e. after the closing brace of a raw data block it would call `nextToken(.whitespace_delimiter_only)` anticipating that it would need that token for the next resource).

This was a problem in certain circumstances, for example `LANGUAGE <num>, <num>` can be a top-level statement or a statement within the optional-statements of another resource (like STRINGTABLE). For the top-level one, the next token would need to be lexed as .whitespace_delimiter_only, but for the STRINGTABLE one, the next token would need to be lexed as .normal.

Instead of trying to anticipate what will come next and lexing a token ahead-of-time, tokens are lexed by the part of the parser than will immediately consume them. If the token is not immediately consumed, a lookahead is done instead so that the token can be re-interpreted/re-lexed in a different context if necessary.