More code page awareness progress
Instead of storing code page in each relevent Node, use an array as a lookup that keeps track of code pages per line number. Also, use a struct that contains both a slice and the code_page (`SourceBytes`) to pass around values that need parsing in order to ensure that we always are able to take the code page into account.