Fix `isRootFile` check and `lookaheadToken`
The assumption that the 'root' file is the first filename added to the source mappings is not a safe one, as the source mapping only adds filenames once it maps a 'real' line, not whenever it sees a `#line` directive. Instead, we manually lookup the input filename in the mappings and then set the root filename offset to the proper value.

This fixes some regressions in the win32-samples-rc-tests, since some #pragma code_page's were being skipped due to the faulty isRootFile checks + lookaheadToken not handling the CodePagePragmaInIncludedFile error correctly.