Fix parseAndRemoveLineCommands for inputs that don't have #line commands
We now run parseAndRemoveLineCommands for all inputs, which allows us to always have correct line numbers regardless of the `no-preprocess` option (this is necessary because we still remove multiline comments) and to work with inputs that have already been run through a preprocessor (i.e. inputs that may have #line commands in them).