cli: Fix invalid option tests
These test cases depended on the first character of the invalid option not itself being a valid option which was invalidated when the /n option was added. Using a _ as the first character of the invalid options should remove the danger of these tests breaking again in the same way.