Initialize source mapping root filename with the initial filename if given
This avoids problems when compiling files that contain only #line directives that are different than the root filename (likely impossible when run through the preprocessor, but possible if /no-preprocess is set). Now, even if there are no references to the initial filename in the mappings, there is still an entry in the table and the root_filename_offset is set automatically.