source_mapping: Fix CRLF line ending handling
Before, it would hit the \r in the switch and then reset the state so the next char would be \n and it would then treat that as a new line. Now, we process line ending pairs together when they exist so that when the state changes any line endings are legitimately part of a new line.