Correct line ending handling within multiline comments
Previously, all line endings would be collapsed when removing comments and replaced with a space if there were any line endings within a multiline comment. Instead, the Win32 preprocessor keeps line endings that are in multiline comments but ignores bare \r, so that's now what is emulated.

However, there is a bug in the clang preprocessor that converts CRLF in multiline comments to CRCRLF which means that the new strategy will lead to line counts in the preprocessed file that do not match the original file, so error message printing may be affected in CRLF files. See https://github.com/llvm/llvm-project/issues/62614