main: Use clang as the preprocessor by default, fallback to zig cc
This will likely be reversed in the future, but there are some things that need to be addressed before zig cc can be the default. The currently blocker is that the 'atlmfc' include directory is not added to the libc include directories by zig cc when targeting MSVC, which is especially crucial when working with Windows resources.