Make zig the default preprocessor + improve auto include path detection
Zig (via zig clang) is now the default preprocessor, with clang as the fallback. `zig clang` is used over `zig cc` because `zig cc` automatically appends options that are unwanted for the purposes of .rc preprocessing (like defining _DEBUG/NDEBUG).

Resinator will now also try to autodetect the system include paths by default, using MSVC if it's available, and falling back to the Zig-distributed MinGW includes (if MSVC is not available and Zig is used as the preprocessor). The particular behavior of include autodetection can be controlled with the `/:auto-includes` CLI option.

This include path autodetection is intended to allow for the default setup to work for most users, even when compiling on non-Windows systems (as long as Zig is used as the preprocessor).