source_mapping: Protect against NUL bytes in #line filenames
This would lead to extremely degraded performance since StringTable relies on NUL-terminated strings, so every lookup would fail since it would compare the slice to the NUL-terminated slice in the StringTable which would be of a different length. Instead, any filenames with a NUL byte within the filename is just ignored as invalid.