Fix potential integer overflow in string literal parsing
When "backtracking" the goal really is for the index to be the same on the next iteration of the loop. So, instead of subtracting in a defer and unconditionally incrementing in the while loop continue expression, we instead just avoid incrementing when "backtracking" to avoid the possible integer overflow in certain rare scenarios.