Improvements and fixes to source mapping and diagnostic line printing
Before, an ArrayList with one element for each line in the file was used for source mapping, with each line in the final input being mapped to a span of lines in the source file. This used more memory than it needed to in the average case (where not many lines have mappings to a span different than its own line number), and contained a bad performance regression in the worst case (when a giant multiline comment was in a file).

This commit makes source mapping use a Treap instead, and makes it so that nodes are only created as-needed when lines need to be mapped.

Outside of performance stuff, things should also now generally agree on where each line is in the file and how many lines there are.