Implement zero-external-dependency automatic system includes
For MinGW, a compressed .tar.zst containing all the MinGW include files is embedded into resinator and is then extracted into a global cache dir on-demand. This adds about ~6MiB to the size of the resinator binary, but makes resinator able to cross-compile from any host out-of-the-box with zero dependencies. See https://github.com/squeek502/compressed_mingw_includes for the details of the compressed header bundle.

For MSVC, include paths are detected using similar logic to Zig's windows_sdk.zig:
- The MSVC installation include path is found by emulating ISetupConfiguration but without using COM
- The latest Windows SDK is found using similar logic that the MSVC 'vcvarsall.bat' batch files use to populate the INCLUDE env var