Use an arena allocator for arocc preprocessing
Since we only care about the preprocessed output and none of the intermediate allocations, using an arena allocator makes perfect sense here and significantly improves the performance of preprocessing.

For example, this takes the ReleaseFast runtime of a particularly expensive `.rc` file (GameDefinition.rc of Windows-classic-samples) from 7.5 seconds down to 560 milliseconds (rc.exe takes 720 milliseconds to compile the same `.rc` file for reference).