bitmaps: Make missing color palette bytes a hard error instead of a warning
Bitmaps with missing color palette bytes are treated as invalid by both the Windows and Firefox bitmap parsers, and therefore emulating the Windows behavior of padding out the missing bytes doesn't make much sense. This simplifies the handling of this edge case and removes the need for a (to-be-configurable) maximum number of missing palette bytes.