cvtres: Exclude DLGINCLUDE resources from the coff object file
cvtres.exe excludes DLGINCLUDE resources, but there's a bug where it will still write the name of the DLGINCLUDE resource (if it's a string) into the Resource Directory Strings section of the data (even though the string(s) in question are not actually referenced by anything, so they're entirely dead weight).

We avoid writing these orphaned strings, and therefore use a hacky workaround in the test code: we force the ID of generated DLGINCLUDE resources to always be an ordinal, and never a string.