cli: Stop parsing options when an argument is inferred to be a filepath
resinator currently can't be used as a drop-in RC compiler for cross-compiling via CMake for two reasons:

- CMake infers the compiler 'type' from the name of the command--it defaults to rc.exe, but switches to using windres options if the command name includes 'windres'
- CMake passes options to rc.exe as `/fo /some/path/to/output.res /some/path/to/input.rc`

Before this commit, resinator would treat the absolute input path as options-to-be-parsed rather than a filepath, and therefore `resinator` would not work-as-expected when used with CMake. After this commit, resinator attempts to infer whether or not the last two arguments are filepaths, and, if so, will stop parsing options.

This is not something that rc.exe has to deal with since it's Windows-only, but resinator does since it's cross-platform but still conforms to the rc.exe CLI interface which allows `/` as an option prefix.