Add `--:auto-includes` option to windres subcommand
This might be a mistake, but the goal is to make the windres subcommand able to compile .rc files with included Windows headers out-of-the-box. This is mostly just a convenience thing, but the default of using the resinator-provided includes is different than the default of windres proper which would require you to specify include paths.

This is definitely worth revisiting and may be removed in the future.