Trick the meson build system into thinking we're rc.exe/GNU windres
Part 1 of this was c093db3b0bed90ae8ca237b310c02aace86da0cb which made the help text get printed to stdout instead of stderr.

This is part 2, where we strategically print the specific strings that meson looks for when determining the type of RC compiler:
- `/?` now includes the string 'Microsoft Resource Compiler'
- The windres subcommand `--version` flag is now supported and the output includes the string 'GNU windres'

With this, `resinator` gets recognized as the Microsoft Resource Compiler, and `resinator windres` gets recognized as GNU windres by meson and everything just works (the benefit of being drop-in CLI compatible).

Closes #17